requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
git-filter-repo==2.47.0
idna==3.11
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
